# docs/openapi.yaml
openapi: 3.0.0
info:
  title: rFactor2 Engineer Backend API
  version: 3.0.0
  description: RESTful API for AI-powered sim-racing telemetry analysis and vehicle setup recommendations. All file parsing is done client-side (WASM); the backend only runs the AI analysis pipeline.
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://telemetria.bot.nu/api
    description: Production server
...
paths:
...
  /analyze_preparsed:
    post:
      tags: [Analysis]
      summary: Analyze client-preparsed telemetry/setup payload
      operationId: analyzePreparsed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channels:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: number
                time_col:
                  type: string
                lap_col:
                  type: string
                session_stats:
                  $ref: '#/components/schemas/SessionStats'
                setup:
                  type: object
                  properties:
                    sections:
                      type: object
                      additionalProperties:
                        type: object
                        properties:
                          name:
                            type: string
                          params:
                            type: object
                            additionalProperties:
                              type: string
                          read_only_params:
                            type: array
                            items:
                              type: string
                model:
                  type: string
                provider:
                  type: string
                  enum: [ollama_cloud]
                ollama_base_url:
                  type: string
                ollama_api_key:
                  type: string
                fixed_params:
                  type: array
                  items:
                    type: string
              required: [channels, setup]
      responses:
        '200':
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
  /analyze_preparsed_stream:
    post:
      tags: [Analysis]
      summary: Analyze client-preparsed telemetry/setup payload with SSE progress events
      operationId: analyzePreparsedStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channels:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: number
                time_col:
                  type: string
                lap_col:
                  type: string
                session_stats:
                  $ref: '#/components/schemas/SessionStats'
                setup:
                  type: object
                model:
                  type: string
                provider:
                  type: string
                  enum: [ollama_cloud]
                ollama_base_url:
                  type: string
                ollama_api_key:
                  type: string
                fixed_params:
                  type: array
                  items:
                    type: string
              required: [channels, setup]
      responses:
        '200':
          description: Server-Sent Events stream ending with an AnalysisResponse result payload
          content:
            text/event-stream:
              schema:
                type: string
  /models:
    get:
      tags: [Models]
      summary: List available LLM models
      operationId: getModels
      parameters:
        - name: ollama_base_url
          in: query
          schema:
            type: string
        - name: ollama_api_key
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
            enum: [ollama_cloud]
      responses:
        '200':
          description: Available models
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      type: string
...
components:
  schemas:
    SetupChange:
      type: object
      properties:
        parameter:
          type: string
        old_value:
          type: string
        new_value:
          type: string
        reason:
          type: string
        change_pct:
          type: string
    AnalysisResponse:
      type: object
      properties:
        driving_analysis:
          type: string
          description: Driving improvement points in Spanish (variable number based on telemetry data)
        setup_analysis:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/SetupChange'
        full_setup:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/SetupChange'
        agent_reports:
          type: array
          items:
            type: object
        setup_agent_reports:
          type: array
          items:
            type: object
        chief_reasoning:
          type: string
        llm_provider:
          type: string
        llm_model:
          type: string
    LapStats:
      type: object
      properties:
        lap:
          type: integer
        duration:
          type: number
        avg_speed:
          type: number
        max_speed:
          type: number
        avg_throttle:
          type: number
        max_throttle:
          type: number
        avg_brake:
          type: number
        max_brake:
          type: number
        avg_rpm:
          type: number
    SessionStats:
      type: object
      properties:
        circuit_name:
          type: string
        total_laps:
          type: integer
        best_lap_time:
          type: number
        avg_lap_time:
          type: number
        laps:
          type: array
          items:
            $ref: '#/components/schemas/LapStats'